+++ components/SignUpForm.module.scss
@@ -1,0 +1,26 @@
+.form {
+  margin-bottom: 3.2rem;
+
+  .input-group {
+    display: flex;
+    flex-direction: column;
+    gap: 2.4rem;
+    margin-bottom: 3rem;
+  }
+
+  label {
+    display: block;
+    margin-bottom: 1.2rem;
+    font-size: 1.4rem;
+  }
+
+  .error-message {
+    margin-top: 0.6rem;
+    font-size: 1.4rem;
+    color: var(--color-red);
+  }
+
+  .button {
+    width: 100%;
+  }
+}
