+++ pages/signup.tsx
@@ -1,48 +1,16 @@
 import Link from 'next/link';
 import Image from 'next/image';
 import Container from '@/components/Container';
-import Button from '@/components/Button';
-import Input from '@/components/Input';
+import SignUpForm from '@/components/SignUpForm';
 import styles from '@/styles/SignupPage.module.scss';
 import logoImg from '@/public/assets/images/logo.svg';
 import googleImg from '@/public/assets/images/google.png';
 import kakaoImg from '@/public/assets/images/kakao.png';
 import classNames from 'classnames/bind';
-import { useState, ChangeEvent, FormEvent } from 'react';
-import { useRouter } from 'next/router';
 
 const cx = classNames.bind(styles);
 
 export default function SignUpPage() {
-  const [values, setValues] = useState({
-    name: '',
-    email: '',
-    password: '',
-    passwordRepeat: '',
-  });
-
-  const router = useRouter();
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-
-    setValues((prevValues) => ({
-      ...prevValues,
-      [name]: value,
-    }));
-  };
-
-  async function handleSubmit(e: FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-
-    try {
-      console.log('회원가입 성공');
-      router.push('/folder');
-    } catch (error) {
-      console.error('Error:', error);
-    }
-  }
-
   return (
     <Container page={'users'}>
       <header className={cx('header')}>
@@ -59,52 +27,7 @@ export default function SignUpPage() {
         </div>
       </header>
 
-      <form className={cx('form')} onSubmit={handleSubmit}>
-        <div className={cx('input-group')}>
-          <div>
-            <label htmlFor="email">이메일</label>
-            <Input
-              id="email"
-              className={cx('input')}
-              name="email"
-              type="text"
-              placeholder="example@email.com"
-              value={values.email}
-              onChange={handleChange}
-            />
-            <p className={cx('error-message')}>이메일 입력해주세요</p>
-          </div>
-          <div>
-            <label htmlFor="password">비밀번호</label>
-            <Input
-              id="password"
-              className={cx('input')}
-              name="password"
-              type="password"
-              placeholder=""
-              value={values.password}
-              onChange={handleChange}
-            />
-            <p className={cx('error-message')}>비밀번호를 입력해주세요</p>
-          </div>
-          <div>
-            <label htmlFor="passwordRepeat">비밀번호 확인</label>
-            <Input
-              id="passwordRepeat"
-              className={cx('input')}
-              name="passwordRepeat"
-              type="password"
-              placeholder=""
-              value={values.passwordRepeat}
-              onChange={handleChange}
-            />
-            <p className={cx('error-message')}>비밀번호를 입력해주세요</p>
-          </div>
-        </div>
-        <Button variant={'primary'} size={'large'} className={cx('button')}>
-          회원가입
-        </Button>
-      </form>
+      <SignUpForm />
 
       <footer className={cx('footer')}>
         <h2 className={cx('title')}>다른 방식으로 가입하기</h2>
